more update WIP

+++ test/test.js
@@ -2,41 +2,59 @@
 
 var fs = require('fs');
 var execSync = require('execSync');
+var expect = require('expect.js');
+var spawn = require('child_process').spawn;
+var _ = require('underscore');
 
-process.chdir('test/fixture');
+describe('Bowinst',function(){
 
-var indexHtml = fs.readFileSync('_index.html',{encoding:'utf8'});
-fs.writeFileSync('index.html',indexHtml,{encoding:'utf8'});
-fs.writeFileSync('custom.html',indexHtml,{encoding:'utf8'});
+    var oldCwd = process.cwd();
 
-var setupJs = fs.readFileSync('js/_setup.js',{encoding:'utf8'});
-fs.writeFileSync('js/setup.js',setupJs,{encoding:'utf8'});
+    before(function(){
 
-fs.unlinkSync('.bowinst.js');
+        process.chdir('test/fixture');
 
-exports.bowinst = {
-    setUp: function(done) {
-        done();
-    },
-    install: function(test) {
-        test.expect(2);
+        var indexHtml = fs.readFileSync('_index.html',{encoding:'utf8'});
+        fs.writeFileSync('index.html',indexHtml,{encoding:'utf8'});
+        fs.writeFileSync('custom.html',indexHtml,{encoding:'utf8'});
 
+        var setupJs = fs.readFileSync('js/_setup.js',{encoding:'utf8'});
+        fs.writeFileSync('js/setup.js',setupJs,{encoding:'utf8'});
+
+        if (fs.existsSync('.bowinst.js')) fs.unlinkSync('.bowinst.js');
+
+        var sampleBowerjson = fs.readFileSync('samplev1_bower.json',{encoding:'utf8'});
+        fs.writeFileSync('bower_components/sample/bower.json',sampleBowerjson,{encoding:'utf8'});  
+
+        var updateHtml = fs.readFileSync('_update.html',{encoding:'utf8'});
+        fs.writeFileSync('update.html',updateHtml,{encoding:'utf8'});             
+
+    });
+
+    after(function(){
+
+        process.chdir(oldCwd);
+
+    });
+
+    it('should install.',function(){
+        
         var result = execSync.exec('../../bin/bowinst install jquery');
         var result = execSync.exec('../../bin/bowinst install angular-unstable');
         var result = execSync.exec('../../bin/bowinst install angular-soft-forms');
 
         var actual = fs.readFileSync('index.html','utf8');
         var expected = fs.readFileSync('../expected/installed.html','utf8');
-        test.equal(actual, expected, 'should install js and css into index.html correctly.');
+        expect(actual).to.eql(expected);
 
         var actual = fs.readFileSync('js/setup.js','utf8');
         var expected = fs.readFileSync('../expected/installed_setup.js','utf8');
-        test.equal(actual, expected, 'should install the Angular module into setup.js correctly.');
 
-        test.done();
-    },
-    uninstall: function(test) {
-        test.expect(2);
+        expect(actual).to.eql(expected);
+
+    });
+
+    it('should uninstall.',function(){
 
         var result = execSync.exec('../../bin/bowinst uninstall jquery');
         var result = execSync.exec('../../bin/bowinst uninstall angular-unstable');
@@ -44,16 +62,45 @@ exports.bowinst = {
 
         var actual = fs.readFileSync('index.html','utf8');
         var expected = fs.readFileSync('_index.html','utf8');
-        test.equal(actual, expected, 'should uninstall js and css from index.html correctly.');
+        expect(actual).to.eql(expected);
 
         var actual = fs.readFileSync('js/setup.js','utf8');
         var expected = fs.readFileSync('js/_setup.js','utf8');
-        test.equal(actual, expected, 'should uninstall the Angular module from setup.js correctly.');
+        expect(actual).to.eql(expected);
 
-        test.done();
-    },
-    customConfig: function(test) {
-        test.expect(1);
+    });
+
+    it('should update (preinstall then install) correctly.',function(done){
+
+        var bowinstjs = fs.readFileSync('_.bowinst_update.js',{encoding:'utf8'});
+        fs.writeFileSync('.bowinst.js',bowinstjs,{encoding:'utf8'});        
+
+        var proc = spawn('../../bin/bowinst', ['preinstall', 'sample'], { env: _.extend({bower_pid:"1234"},process.env) });
+
+        proc.on('close', function (code) {
+            
+            var bowerjsonv2 = fs.readFileSync('samplev2_bower.json');
+            fs.writeFileSync('bower_components/sample/bower.json',bowerjsonv2);
+
+            proc = spawn('../../bin/bowinst', ['install', 'sample'], { env: _.extend({bower_pid:"1234"},process.env) });
+
+            proc.on('close',function(code){
+
+                var actual = fs.readFileSync('update.html','utf8');
+                var expected = fs.readFileSync('../expected/updated.html','utf8');
+                expect(actual).to.eql(expected);
+
+                var actual = fs.readFileSync('js/setup.js','utf8');
+                var expected = fs.readFileSync('../expected/updated_setup.js','utf8');
+                expect(actual).to.eql(expected);
+
+                done();
+            });
+        });
+
+    });
+
+    it('should use a custom config correctly.',function(){
 
         var bowinstjs = fs.readFileSync('_.bowinst.js',{encoding:'utf8'});
         fs.writeFileSync('.bowinst.js',bowinstjs,{encoding:'utf8'});
@@ -62,9 +109,8 @@ exports.bowinst = {
 
         var actual = fs.readFileSync('custom.html','utf8');
         var expected = fs.readFileSync('../expected/installed_custom.html','utf8');
-        test.equal(actual, expected, 'should use custom configuration successfully.');
+        expect(actual).to.eql(expected);
 
-        test.done();        
+    });
 
-    }
-};
+});
